# docs/api-contract.yaml
openapi: 3.0.3
info:
  title: Semantic Business Map API
  version: 1.0.0
  description: Evidence-first application contract. Source for generated frontend types. No graph snippets are returned by the canvas endpoint.
servers:
  - url: /api/v1
paths:
...
  /projects/{projectId}/repositories:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
    post:
      operationId: addRepository
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RepositoryInput' }
      responses:
        '201':
          description: Attached repository
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Repository' }
...
  /semantic/nodes/{nodeId}/reviews:
    parameters:
      - $ref: '#/components/parameters/NodeId'
    post:
      operationId: reviewNode
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReviewInput' }
      responses:
        '200': { description: Review recorded }
        '201': { description: Review recorded }
        '204': { description: Review recorded }
...
components:
  parameters:
    ProjectId:
      in: path
      name: projectId
      required: true
      schema: { type: string }
...
    NodeId:
      in: path
      name: nodeId
      required: true
      schema: { type: string }
...
  schemas:
...
    RepositoryInput:
      type: object
      required: [url, branch]
      properties:
        url: { type: string, minLength: 1 }
        branch: { type: string, minLength: 1 }
    Repository:
      allOf:
        - $ref: '#/components/schemas/RepositoryInput'
        - type: object
          required: [id]
          properties:
            id: { type: string }
...
    ReviewInput:
      type: object
      required: [decision, comment, editedTitle, editedDescription]
      properties:
        decision:
          type: string
          enum: [CONFIRMED, REJECTED, EDITED, MERGED, MARKED_TECHNICAL, NEEDS_REVIEW]
        comment: { type: string }
        editedTitle: { type: string }
        editedDescription: { type: string }
        mergeTargetId: { type: string, format: uuid }
...